# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ecommerce API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users, admin]
      summary: List users (admin)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create / register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/addresses:
    get:
      operationId: listUserAddresses
      tags: [users]
      summary: List user addresses
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createUserAddress
      tags: [users]
      summary: Add user address
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Created address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

  # ─── Reviews ──────────────────────────────────────────────────────────────
...
  /users/{id}/wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get user wishlist
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add product to wishlist
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistItemInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /users/{id}/wishlist/{productId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove product from wishlist
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed

  # ─── Analytics / Admin ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, slug, price, images, categoryId, inStock, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        compareAtPrice: { type: ["number", "null"] }
        images: { type: array, items: { type: string } }
        categoryId: { type: integer }
        categoryName: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        variants: { type: array, items: { $ref: "#/components/schemas/ProductVariant" } }
        inStock: { type: boolean }
        stockCount: { type: ["integer", "null"] }
        isFeatured: { type: boolean }
        averageRating: { type: ["number", "null"] }
        reviewCount: { type: integer }
        createdAt: { type: string }
    ProductVariant:
      type: object
      required: [id, name, value]
      properties:
        id: { type: integer }
        name: { type: string }
        value: { type: string }
        priceModifier: { type: ["number", "null"] }
        stockCount: { type: ["integer", "null"] }
...
    User:
      type: object
      required: [id, email, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        firstName: { type: ["string", "null"] }
        lastName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        role: { type: string, enum: [customer, admin] }
        orderCount: { type: integer }
        totalSpent: { type: number }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        password: { type: string }
    UserUpdate:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        avatarUrl: { type: string }
    UserListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/User" } }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Address ──
    Address:
      type: object
      required: [id, line1, city, state, postalCode, country]
      properties:
        id: { type: integer }
        line1: { type: string }
        line2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
    AddressInput:
      type: object
      required: [line1, city, state, postalCode, country]
      properties:
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }

    # ── Review ──
...
    WishlistItem:
      type: object
      required: [id, userId, product, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        product: { $ref: "#/components/schemas/Product" }
        createdAt: { type: string }
    WishlistItemInput:
      type: object
      required: [productId]
      properties:
        productId: { type: integer }

    # ── Analytics ──
...