# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ecommerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  avatarUrl?: string;
}